requirements.txt
minio==7.1.*
boto3==1.28.*
python-dotenv
